PointAndClick: Added color change on hover events for buttons

+++ PointAndClick/Code/Game/Map.cpp
@@ -127,7 +127,7 @@ void Map::CenterCameraOnPlayer() const
 void Map::Render() const
 {
 	std::vector<Vertex_PCU> vertexes;
-	AppendVertsForAABB2D( vertexes, AABB2( Vec2::ZERO, Vec2( m_width, m_height ) ), Rgba8::WHITE );
+	AppendVertsForAABB2D( vertexes, AABB2( Vec2::ZERO, Vec2( (float)m_width, (float)m_height ) ), Rgba8::WHITE );
 
 	g_renderer->BindTexture( 0, m_mapDef->m_backgroundTexture );
 	g_renderer->DrawVertexArray( vertexes );
